Handle concurrency for epics and user-stories. Add column to epics table saving user_id_editing to check if epic or user-stories is available to edit or not

+++ app/Http/Controllers/UserStoryController.php
@@ -25,6 +25,34 @@ class UserStoryController
     ]);
   }
 
+  public function edit(Request $request, UserStory $userStory)
+  {
+    $project = $userStory->get_project_of_user_story();
+
+    if(!$request->user()->has_basic_permissions_to_project($project)){
+      return response()->json([
+          'success' => false,
+          'message' => 'Not authorized'
+       ], 403);
+    }
+    
+    if(!$userStory->isAvailableToEdit()){
+      return response()->json([
+          'success' => false,
+          'message' => 'Not available to edit now. Try later'
+       ], 422);
+    };
+
+    $epic = $userStory->epic()->first();
+    $epic->user_id_editing = $request->user()->id;
+    $epic->save();
+
+    return response()->json([
+        'success' => true,
+        'userStory' => $userStory
+    ]);
+  }
+
   public function update(UserStoryRequest $request, UserStory $userStory)
   {
     $project = $userStory->get_project_of_user_story();
@@ -36,7 +64,19 @@ class UserStoryController
        ], 403);
     }
 
+    $epic = $userStory->epic()->first();
+
+    if($epic->user_id_editing != $request->user()->id){
+      return response()->json([
+        'success' => false,
+        'message' => 'Not available to edit now. Try later'
+     ], 422);
+    };
+
     $userStory->update($request->all());
+    
+    $epic->user_id_editing = null;
+    $epic->save();
   
     return response()->json([
         'success' => true,
@@ -54,6 +94,13 @@ class UserStoryController
           'message' => 'Not authorized'
        ], 403);
     }
+
+    if(!$userStory->isAvailableToEdit()){
+      return response()->json([
+          'success' => false,
+          'message' => 'Not available to edit now. Try later'
+       ], 422);
+    };
 
     $userStory->delete();
